replace folders functionality to popup.js #1
